# OpenAPI/OpenAPI.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Spring boot template",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/healthcheck": {
      "get": {
        "tags": [
          "healthcheck"
        ],
        "description": "Healthcheck endpoint",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/poi" : {
      "post" : {
        "operationId": "addPoi",
        "tags" : [
          "poi"
        ],
        "description": "Insert poi",
        "parameters": [
          {
            "$ref" : "common/params/headers/accountIDParam.json"
          }
        ],
        "requestBody" : {
          "$ref" : "APIspecs/requests/poiPostRequestBody.json"
        },
        "responses": {
          "201" : {
            "$ref" : "APIspecs/responses/poiCreated.json"
          },
          "400": {
            "$ref": "APIspecs/responses/badRequest.json"
          },
          "409": {
            "$ref": "APIspecs/responses/conflict.json"
          }
        }
      }
    },
    "/api/v1/poi/{poi-id}" : {
      "get" : {
        "operationId" : "getPoi",
        "tags" : [
          "poi"
        ],
        "description" : "get Poi",
        "parameters" : [
          {
            "$ref" : "common/params/headers/accountIDParam.json"
          },
          {
            "$ref" : "common/params/pathParams/poiIDParam.json"
          }
        ],
        "responses" : {
          "200" : {
            "$ref": "APIspecs/responses/poiGetOk.json"
          },
          "404" : {
            "$ref": "APIspecs/responses/notFound.json"
          }
        }
      },
      "put" : {
        "operationId" : "updatePoi",
        "tags" : [
          "poi"
        ],
        "description" : "update Poi",
        "parameters" : [
          {
            "$ref" : "common/params/headers/accountIDParam.json"
          },
          {
            "$ref" : "common/params/pathParams/poiIDParam.json"
          }
        ],
        "requestBody" : {
          "$ref" : "APIspecs/requests/poiPutRequestBody.json"
        },
        "responses" : {
          "200" : {
            "$ref": "APIspecs/responses/poiUpdated.json"
          },
          "400": {
            "$ref": "APIspecs/responses/badRequest.json"
          },
          "404" : {
            "$ref": "APIspecs/responses/notFound.json"
          }
        }
      }
    },
    "/api/v1/poi/distance/{meters}" : {
      "get" : {
        "operationId" : "getNearestPois",
        "tags" : [
          "poi"
        ],
        "description" : "get nearest Pois",
        "parameters" : [
          {
            "$ref" : "common/params/headers/accountIDParam.json"
          },
          {
            "$ref" : "common/params/pathParams/metersParam.json"
          },
          {
            "$ref" : "common/params/queryParams/latitudeQueryParam.json"
          },
          {
            "$ref" : "common/params/queryParams/longitudeQueryParam.json"
          },
          {
            "$ref" : "common/params/queryParams/pageQueryParam.json"
          },
          {
            "$ref" : "common/params/queryParams/pageSizeQueryParam.json"
          }
        ],
        "responses" : {
          "200" : {
            "$ref": "APIspecs/responses/searchNearestOk.json"
          },
          "400" : {
            "$ref": "APIspecs/responses/badRequest.json"
          }
        }
      }
    }
  }
}